refactor intro screen from being a screen to just a component
