+++ test/RootStateSchema.json
@@ -3316,7 +3316,6 @@
                 "ImportWallet",
                 "Invite",
                 "JumpstartEnterAmount",
-                "JumpstartIntroScreen",
                 "JumpstartSendConfirmation",
                 "JumpstartShareLink",
                 "JumpstartTransactionDetailsScreen",
